# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Vantage Mock Interview API
  version: 1.0.0
  description: >
    MVP REST contract for the mock interview preparation agent. Memory endpoints are
    authenticated and derive the acting user from the session cookie rather than the request body.
    Mentor queue, mentor intervention, and mentor-scoped realtime endpoints require a mentor or
    admin role.
servers:
  - url: http://localhost:3000
    description: Default local development URL. The actual dev port may move if 3000 is already in use.
paths:
...
  /events/stream:
    get:
      summary: Open a realtime Server-Sent Events stream for session or mentor updates
      parameters:
        - in: query
          name: scope
          required: false
          schema:
            type: string
            enum: [session, mentor]
        - in: query
          name: session_id
          required: false
          schema:
            type:
              - string
              - 'null'
      responses:
        '200':
          description: SSE stream of guardrail and mentor events
...